Move update_available flag to MainViewController; fix hero image directory URL
The update_available flag was a loose static in cg_ui.c with a destructive
read getter (Cg_IsUpdateAvailable), exported as a public API just for
HomeViewController to call. This created an unusual temporal coupling and
a side-effecting getter anti-pattern.

Move the flag to MainViewController.updateAvailable, which already owns the
UI lifecycle and already handles similar conditional navigation in its own
viewWillAppear (teams check). Cg_UpdateInstaller now sets the field directly.
The dialog is presented in MainViewController::viewWillAppear alongside the
existing teams logic. HomeViewController::viewWillAppear is no longer needed.

Also fix fetchHeroImages skipping the S3 directory entry: the XML listing
includes <Key>hero-images/</Key> (the directory itself), which after stripping
the prefix produces an empty filename and a URL equal to the listing endpoint,
causing a spurious warning on every startup.
