+++ src/cgame/default/ui/home/HomeViewController.c
@@ -23,20 +23,10 @@
 
 #include "HomeViewController.h"
 
-#include "DialogViewController.h"
 #include "LeaderboardViewController.h"
 #include "StatsViewController.h"
 
 #define _Class _HomeViewController
-
-#pragma mark - Delegates
-
-/**
- * @brief Dialog::okFunction for opening the releases page.
- */
-static void openReleasesPage(ident data) {
-  SDL_OpenURL(QUETOO_RELEASES_URL);
-}
 
 #pragma mark - Object
 
@@ -87,27 +77,6 @@ static void loadView(ViewController *self) {
 }
 
 /**
- * @see ViewController::viewWillAppear(ViewController *)
- */
-static void viewWillAppear(ViewController *self) {
-
-  super(ViewController, self, viewWillAppear);
-
-  if (Cg_IsUpdateAvailable()) {
-    const Dialog dialog = {
-      .message = "A new version of Quetoo is available. Download now?",
-      .ok = "Yes",
-      .cancel = "No",
-      .okFunction = openReleasesPage
-    };
-
-    ViewController *viewController = (ViewController *) $(alloc(DialogViewController), initWithDialog, &dialog);
-    $(self, addChildViewController, viewController);
-    release(viewController);
-  }
-}
-
-/**
  * @see Class::initialize(Class *)
  */
 static void initialize(Class *clazz) {
@@ -115,7 +84,6 @@ static void initialize(Class *clazz) {
   ((ObjectInterface *) clazz->interface)->dealloc = dealloc;
 
   ((ViewControllerInterface *) clazz->interface)->loadView = loadView;
-  ((ViewControllerInterface *) clazz->interface)->viewWillAppear = viewWillAppear;
 }
 
 /**
