+++ src/cgame/default/ui/main/MainViewController.c
@@ -64,6 +64,13 @@ static void didClickNavigateViewController(Button *button) {
 }
 
 /**
+ * @brief Opens the releases page.
+ */
+static void openReleasesPage(ident data) {
+  SDL_OpenURL(QUETOO_RELEASES_URL);
+}
+
+/**
  * @brief Quit the game.
  */
 static void quit(ident data) {
@@ -210,6 +217,21 @@ static void viewWillAppear(ViewController *self) {
     if (team == TEAM_NONE) {
       $(this, navigateToViewController, _TeamsViewController());
     }
+  }
+
+  if (this->updateAvailable) {
+    this->updateAvailable = false;
+
+    const Dialog dialog = {
+      .message = "A new version of Quetoo is available. Download now?",
+      .ok = "Yes",
+      .cancel = "No",
+      .okFunction = openReleasesPage
+    };
+
+    ViewController *viewController = (ViewController *) $(alloc(DialogViewController), initWithDialog, &dialog);
+    $(self, addChildViewController, viewController);
+    release(viewController);
   }
 }
